:truck: Move files to src
