+++ src/components/SEO.tsx
@@ -5,7 +5,7 @@
 import Head from 'next/head'
 import { useRouter } from 'next/router'
 
-import siteMetadata from '@data/siteMetadata'
+import siteMetadata from '@/data/siteMetadata'
 
 type TCommonSEOProps = {
   title: string
